requirements.txt
robotframework==7.0
robotframework-browser==18.0.0
robotframework-seleniumlibrary==6.2.0
robotframework-requests==0.9.6
robotframework-databaselibrary==1.2.4
robotframework-appiumlibrary==2.0.0
robotframework-pabot==2.18.0
robotframework-excellib==3.0.0
robotframework-jsonlibrary==0.5
allure-robotframework==2.13.5
selenium==4.16.0
playwright==1.40.0
appium-python-client==3.1.0
requests==2.31.0
PyMySQL==1.1.0
faker==22.0.0
python-dotenv==1.0.0
pytest==7.4.3
opencv-python==4.8.1
pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
pyyaml==6.0.1
jsonschema==4.20.0
beautifulsoup4==4.12.2
lxml==4.9.3
paramiko==3.4.0
cryptography==41.0.7
scipy==1.11.4
numpy==1.26.2
tensorflow-lite==2.14.0